refactor: move ArrayTool to Json
